% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsAreasClimaticasList.R
\name{getwsAreasClimaticasList}
\alias{getwsAreasClimaticasList}
\title{Obtener la lista de áreas climáticas de la base de datos del CLIMA}
\usage{
getwsAreasClimaticasList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{CODIGO}, \code{DENOMINACION}, \code{PKACL}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de las áreas climáticas en la base de datos CLIMA.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
areas_climaticas <- getwsAreasClimaticasList(idsesion)
}

}
