% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsAreasGeograficasList.R
\name{getwsAreasGeograficasList}
\alias{getwsAreasGeograficasList}
\title{Obtener la lista de áreas geográficas de la base de datos del CLIMA}
\usage{
getwsAreasGeograficasList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{CODIGO}, \code{DENOMINACION}, \code{PKAGE}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de las áreas geográficas en la base de datos CLIMA.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
areas_geograficas <- getwsAreasGeograficasList(idsesion)

}
