% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsComarcasList.R
\name{getwsComarcasList}
\alias{getwsComarcasList}
\title{Obtener la lista de comarcas de la base de datos del CLIMA}
\usage{
getwsComarcasList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{AMBITO}, \code{CCOMARCA}, \code{DENOMINACION}, \code{PKCOM}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de las comarcas en la base de datos CLIMA.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
comarcas <- getwsComarcasList(idsesion)
}

}
