% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosAnuales.R
\name{getwsDatosAnuales}
\alias{getwsDatosAnuales}
\title{Obtener datos anuales a partir de los identificadores de estaciones,
variables y fechas y devolverlos como data.frame}
\usage{
getwsDatosAnuales(idsesion, cestaciones, cvariables, fechas)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{cestaciones}{Character vector. Vector con los identificadores de las estaciones.}

\item{cvariables}{Character vector. Vector con los identificadores de las variables.}

\item{fechas}{Character vector. Vector con las fechas del último día de cada año en formato DD/MM/AAAA.}
}
\value{
data.frame con columnas: cestacion, cvariable, fecha, valor.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la
Junta de Andalucía para obtener datos anuales de varias estaciones, variables
y fechas, y devolverlos como data.frame.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
cestaciones <- '5402'
cvariables <- 'TA1'
fechas <- c('31/12/2000','31/12/2001','31/12/2002','31/12/2003','31/12/2004','31/12/2005')
datos_anuales <- getwsDatosAnuales(idsesion,cestaciones,cvariables,fechas)
head(datos_anuales)
}

}
