% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosAnualesRaw.R
\name{getwsDatosAnualesRaw}
\alias{getwsDatosAnualesRaw}
\title{Obtener datos anuales a partir de los identificadores internos de
estaciones, variables y fechas y devolverlos como data.frame}
\usage{
getwsDatosAnualesRaw(idsesion, pkests, pkvars, pkfecs)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{pkests}{Character vector. Vector con los identificadores internos de las estaciones.}

\item{pkvars}{Character vector. Vector con los identificadores internos de las variables.}

\item{pkfecs}{Character vector. Vector con las identificadores internos de las fechas.}
}
\value{
data.frame con columnas: pkest, pkvar, pkfec, valor.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la
Junta de Andalucía para obtener datos anuales de varias estaciones, variables
y fechas, a partir de sus identificadores internos, y devolverlos como data.frame.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
pkests <- c('2306') # 5402
pkvars <- c('41') # TA1
pkfecs <- c('93502') # 31/12/2020
datos_anuales <- getwsDatosAnualesRaw(idsesion,pkests,pkvars,pkfecs)
head(datos_anuales)
}

}
