% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosDiarios.R
\name{getwsDatosDiarios}
\alias{getwsDatosDiarios}
\title{Obtener datos diarios de una estación, variable y fecha y devolverlos
como data.frame}
\usage{
getwsDatosDiarios(idsesion, cestacion, cvariable, fecha)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{cestacion}{Character. Identificador de la estación.}

\item{cvariable}{Character. Identificador de la variable.}

\item{fecha}{Character. Fecha en formato DD/MM/AAAA.}
}
\value{
data.frame con columnas: cestacion, cvariable, fecha, valor.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la
Junta de Andalucía para obtener un dato diario de una estación, variable y
fecha, y devuelve un data.frame.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
cestacion <- 'EARM22'
cvariable <- 'TD1'
fecha <- '26/10/2025'
result_diaria <- getwsDatosDiarios(idsesion,cestacion,cvariable,fecha)
print(result_diaria)
}

}
