% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosDiarios2.R
\name{getwsDatosDiarios2}
\alias{getwsDatosDiarios2}
\title{Obtener datos diarios de un periodo a partir de los identificadores
de estaciones, variables y fechas, y devolverlos como data.frame}
\usage{
getwsDatosDiarios2(idsesion, cestaciones, cvariables, fecha_ini, fecha_fin)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{cestaciones}{Character vector. Identificadores de las estaciones.}

\item{cvariables}{Character vector. Identificadores de las variables.}

\item{fecha_ini}{Character. Fecha inicial en formato DD/MM/AAAA.}

\item{fecha_fin}{Character. Fecha final en formato DD/MM/AAAA.}
}
\value{
data.frame con columnas: cestacion, cvariable, fecha, valor.
}
\description{
Llama al método \code{getDatosDiarios2} del servicio CLIMA de la Junta de
Andalucía para obtener datos diarios de varias estaciones y variables entre
dos fechas. Las fechas se generan secuencialmente entre \code{fecha_ini} y
\code{fecha_fin} por lo que, en caso de lagunas, pueden no ser correctos.

Se recomienda, en caso de que haya lagunas en la serie, consultar los datos
conjuntamente con los de otra estación cuya serie sí esté completa lo que
garantiza que la respuesta del servicio contiene todos los nodos y deja como
NA los datos ausentes.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
cestaciones <- c('EARM22','SIVA40')
cvariables <- c('TD1','PD23')
fecha_ini <- '15/10/2025'
fecha_fin <- '26/10/2025'
result_diarios_intervalo <- getwsDatosDiarios2(idsesion,cestaciones,cvariables,fecha_ini,fecha_fin)
head(result_diarios_intervalo)
}

}
