% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosDiarios2Raw.R
\name{getwsDatosDiarios2Raw}
\alias{getwsDatosDiarios2Raw}
\title{Obtener datos diarios de un periodo a partir de los identificadores
internos de estaciones, variables y fechas, y devolverlos como data.frame}
\usage{
getwsDatosDiarios2Raw(idsesion, pkests, pkvars, pkfec_ini, pkfec_fin)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{pkests}{Character vector. Identificadores internos de las estaciones.}

\item{pkvars}{Character vector. Identificadores internos de las variables.}

\item{pkfec_ini}{Character. Identificador interno de la fecha inicial.}

\item{pkfec_fin}{Character. Identificador interno de la fecha final.}
}
\value{
data.frame con columnas: pkest, pkvar, pkfec, valor.
}
\description{
Llama al método \code{getDatosDiarios2} del servicio CLIMA de la Junta de
Andalucía para obtener datos diarios de varias estaciones y variables entre
dos PKFEC (identificadores numéricos de fecha). Los PKFEC, se generan
secuencialmente entre \code{pkfec_ini} y \code{pkfec_fin} por lo que, en
caso de lagunas, pueden no ser correctos.

Se recomienda, en caso de que haya lagunas en la serie, consultar los datos
conjuntamente con los de otra estación cuya serie sí esté completa lo que
garantiza que la respuesta del servicio contiene todos los nodos y deja como
NA los datos ausentes.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
pkests <- c("183","143")
pkvars <- c('2','711') # TD1 y PD23
pkfec_ini <- '100730' # 15/10/2025
pkfec_fin <- '100741' # 26/10/2025
result_diarios <- getwsDatosDiarios2Raw(idsesion,pkests,pkvars,pkfec_ini,pkfec_fin)
head(result_diarios)
}

}
