% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosDiariosMulti.R
\name{getwsDatosDiariosMulti}
\alias{getwsDatosDiariosMulti}
\title{Obtener datos diarios a partir de los identificadores de las
estaciones, variables y fechas y devolverlos como data.frame (multiconsulta)}
\usage{
getwsDatosDiariosMulti(idsesion, cestaciones, cvariables, fechas)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{cestaciones}{Character vector. Identificadores de las estaciones.}

\item{cvariables}{Character vector. Identificadores de las variables.}

\item{fechas}{Character vector. Fechas en formato DD/MM/AAAA.}
}
\value{
data.frame con columnas: cestacion, vbariable, fecha y valor
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de
Andalucía para obtener datos diarios de varias estaciones, variables y fechas,
a partir de sus identificadores, devolviendo un data.frame
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
cestaciones <- 'EARM22'
cvariables <- c('TD1','PD23')
fechas <- c('26/10/2025','27/10/2025')
result_diaria <- getwsDatosDiariosMulti(idsesion,pkests,pkvars,pkfecs)
print(result_diaria)
}

}
