% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosDiariosMultiRaw.R
\name{getwsDatosDiariosMultiRaw}
\alias{getwsDatosDiariosMultiRaw}
\title{Obtener datos diarios a partir de los identificadores internos de las
estaciones, variables y fechas y devolverlos como data.frame (multiconsulta)}
\usage{
getwsDatosDiariosMultiRaw(idsesion, pkests, pkvars, pkfecs)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{pkests}{Character vector. Identificadores internos de las estaciones.}

\item{pkvars}{Character vector. Identificadores internos de las variables.}

\item{pkfecs}{Character vector. Identificadores internos de las fechas.}
}
\value{
data.frame con columnas: pkest, pkvar, pkfec y valor
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de
Andalucía para obtener datos diarios de varias estaciones, variables y fechas,
a partir de sus identificadores internos, devolviendo un data.frame
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
pkests <- c('183') # EARM22
pkvars <- c('2','711') # TD1 y PD23
pkfecs <- c('100741','100742') # 26/10/2025 y 27/10/2025
result_diaria <- getwsDatosDiariosMultiRaw(idsesion,pkests,pkvars,pkfecs)
head(result_diaria)
}

}
