% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosDiariosRaw.R
\name{getwsDatosDiariosRaw}
\alias{getwsDatosDiariosRaw}
\title{Obtener un dato diario a partir de los identificadores internos de la
estación, variable y fecha y devolverlos como data.frame}
\usage{
getwsDatosDiariosRaw(idsesion, pkest, pkvar, pkfec)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{pkest}{Character. Identificador interno de la estación.}

\item{pkvar}{Character. Identificador interno de la variable.}

\item{pkfec}{Character. Identificador interno de la fecha.}
}
\value{
data.frame con columnas: pkest, pkvar, pkfec, valor.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la
Junta de Andalucía para obtener un dato diario a partir de los identificadores
internos de la estación, la variable y la fecha, y lo devuelve en un data.frame.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
pkest <- '183' # EARM22
pkvar <- '2' # TD1
pkfec <- '100741' # 26/10/2025
result_diaria <- getwsDatosDiariosRaw(idsesion,pkest,pkvar,pkfec)
print(result_diaria)

}
