% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosExisteEstacionesList.R
\name{getwsDatosExisteEstacionesList}
\alias{getwsDatosExisteEstacionesList}
\title{Obtener la lista de estaciones con datos existentes (interpretado)}
\usage{
getwsDatosExisteEstacionesList(idsesion, cestacion = NULL)
}
\arguments{
\item{idsesion}{\code{character}. Identificador de sesión válido obtenido mediante \code{getwsIDSesion()}.}

\item{cestacion}{\code{character}. Código de estación (opcional, por defecto todas).}
}
\value{
Un \code{data.frame} con las columnas devueltas por el servicio (por ejemplo: \code{ESTPKEST}, \code{FECFINDATOS}, \code{FECHACARGA}, \code{FECINICIODATOS}, \code{LEXISTE}, \code{PKDEE}
}
\description{
Consulta el servicio web CLIMA de la Junta de Andalucía mediante el método
\code{getDatosExisteEstacionesList}, que devuelve la lista de estaciones con datos
existentes según los parámetros de búsqueda. Si se introduce el código de la
estación, devuelve sólo el registro solicitado.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
cestacion <- 'EARM22'
existencias_estaciones <- getwsDatosExisteEstacionesList(idsesion,cestacion)
print(existencias_estaciones)
}

}
