% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosExisteEstacionesListRaw.R
\name{getwsDatosExisteEstacionesListRaw}
\alias{getwsDatosExisteEstacionesListRaw}
\title{Obtener la lista de estaciones con datos existentes (en bruto)}
\usage{
getwsDatosExisteEstacionesListRaw(idsesion, pkest = NULL)
}
\arguments{
\item{idsesion}{\code{character}. Identificador de sesión válido obtenido mediante \code{getwsIDSesion()}.}

\item{pkest}{\code{character}. Código de estación (opcional, por defecto todas).}
}
\value{
Un \code{data.frame} con las columnas devueltas por el servicio (por ejemplo: \code{ESTPKEST}, \code{FECFINDATOS}, \code{FECHACARGA}, \code{FECINICIODATOS}, \code{LEXISTE}, \code{PKDEE}
}
\description{
Consulta el servicio web CLIMA de la Junta de Andalucía mediante el método
\code{getDatosExisteEstacionesList}, que devuelve la lista de estaciones con datos
existentes según los parámetros de búsqueda. Si se introduce el identificador
interno de una estación, devuelve sólo el registro solicitado
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
pkest <- '2306'
existencias_estaciones <- getwsDatosExisteEstacionesListRaw(idsesion,pkest)
print(existencias_estaciones)

}
