% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosIntradiarios.R
\name{getwsDatosIntradiarios}
\alias{getwsDatosIntradiarios}
\title{Obtener datos intradiarios de una estación, variable y fecha y devolverlos
como data.frame}
\usage{
getwsDatosIntradiarios(idsesion, cestacion, cvariable, fecha)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{cestacion}{Character. Identificador de la estación a consultar.}

\item{cvariable}{Character. Identificador de la variable a consultar.}

\item{fecha}{Character. Fecha en formato DD/MM/AAAA.}
}
\value{
data.frame con columnas: código de estación, variable, fecha, hora y valor.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la
Junta de Andalucía para obtener datos intradiarios de una estación, variable y
fecha, y devuelve un data.frame.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
cestacion <- 'EARM22'
cvariable <- 'TI1'
fecha <- '26/10/2025'
datos_intra <- getwsDatosIntradiarios(idsesion,cestacion,cvariable,fecha)
head(datos_intra)
}

}
