% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosIntradiariosMulti.R
\name{getwsDatosIntradiariosMulti}
\alias{getwsDatosIntradiariosMulti}
\title{Obtener datos intradiarios de varias estaciones, variables y/o fechas
y devolverlos como data.frame (multiconsulta)}
\usage{
getwsDatosIntradiariosMulti(idsesion, cestaciones, cvariables, fechas)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{cestaciones}{Character. Vector con los identificadores de las estaciones a consultar.}

\item{cvariables}{Character vector. Vector con los identificadores de las variables a consultar.}

\item{fechas}{Character vector. Vector con las fechas a consultar en formato DD/MM/AAAA.}
}
\value{
data.frame con columnas: cestacion, cvariable, fecha, hora y valor.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de
Andalucía para obtener datos intradiarios de estaciones, variables y
fechas y devuelve un data.frame. Permite combinar varias elementos, manteniendo
siempre un elemento único de partida, por ejemplo, varias estaciones y fechas
para una variable o varias fechas y variables para una estación.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
cestaciones <- c('EARM22')
cvariables <- c('TI1','PI1')
fechas <- c('26/10/2025','27/10/2025')
datos_intra <- getwsDatosIntradiariosMultiRaw(idsesion,cestaciones,cvariables,fechas)
head(datos_intra)
}

}
