% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosIntradiariosMultiRaw.R
\name{getwsDatosIntradiariosMultiRaw}
\alias{getwsDatosIntradiariosMultiRaw}
\title{Obtener datos intradiarios a partir de los identificadores internos de
la estación, la variable y la fecha y devolverlos como data.frame (multiconsulta)}
\usage{
getwsDatosIntradiariosMultiRaw(idsesion, pkests, pkvars, pkfecs)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{pkests}{Character vector. Vector con los identificadores internos de las de las estaciones a consultar.}

\item{pkvars}{Character vector. Vector con los identificadores internos de las variables a consultar.}

\item{pkfecs}{Character vector. Vector con los identificadores internos de las fechas a consultar.}
}
\value{
data.frame con columnas: pkest, pkvar, pkfec, minuto, valor.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de
Andalucía para obtener datos intradiarios de una estación, variable y
fecha a partir de sus identificadores internos, y devuelve un data.frame.
Permite combinar varias variables y fechas para una estación única.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
pkests <- c('183') # EARM22
pkvars <- c('1','63') # TI1 y PI1
pkfecs <- c('100741','100742') # 26/10/2025 y 27/10/2025
datos_intra_raw <- getwsDatosIntradiariosMultiRaw(idsesion,pkests,pkvars,pkfecs)
head(datos_intra_raw)
}

}
