% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosIntradiariosRaw.R
\name{getwsDatosIntradiariosRaw}
\alias{getwsDatosIntradiariosRaw}
\title{Obtener datos intradiarios a partir de los identificadores internos de
la estación, la variable y la fecha y devolverlos como data.frame}
\usage{
getwsDatosIntradiariosRaw(idsesion, pkest, pkvar, pkfec)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{pkest}{Character. Identificador interno de la estación a consultar.}

\item{pkvar}{Character. Identificador interno de la variable a consultar.}

\item{pkfec}{Character. Identificador interno de la fecha a consultar.}
}
\value{
data.frame con columnas: id de estación, id de variable, id de fecha y valor.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de
Andalucía para obtener datos intradiarios de una estación, variable y
fecha a partir de sus identificadores internos, y devuelve un data.frame.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
pkest <- '183' # EARM22
pkvar <- '1' # TI1
pkfec <- '100741' # 26/10/2025
datos_intra <- getwsDatosIntradiariosRaw(idsesion,pkest,pkvar,pkfec)
head(datos_intra)

}
