% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosMensuales.R
\name{getwsDatosMensuales}
\alias{getwsDatosMensuales}
\title{Obtener datos mensuales a partir de los identificadores de estaciones,
variables y fechas y devolverlos como data.frame}
\usage{
getwsDatosMensuales(idsesion, cestaciones, cvariables, fechas)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{cestaciones}{Character vector. Vector con los identificadores de las estaciones.}

\item{cvariables}{Character vector. Vector con los identificadores de las variables.}

\item{fechas}{Character vector. Vector con las fechas del último día de cada mes en formato DD/MM/AAAA.}
}
\value{
data.frame con columnas: cestacion, cvariable, fecha, valor.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la
Junta de Andalucía para obtener datos mensuales de varias estaciones, variables
y fechas, y devolverlos como data.frame.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
cestaciones <- c('EARM22','EARM01')
cvariables <- c('TM1')
fechas <- c('31/12/2020')
datos_mensuales <- getwsDatosMensuales(idsesion,cestaciones,cvariables,fechas)
head(datos_mensuales)
}

}
