% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsDatosMensualesRaw.R
\name{getwsDatosMensualesRaw}
\alias{getwsDatosMensualesRaw}
\title{Obtener datos mensuales a partir de los identificadores internos de
estaciones, variables y fechas y devolverlos como data.frame}
\usage{
getwsDatosMensualesRaw(idsesion, pkests, pkvars, pkfecs)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido obtenido previamente.}

\item{pkests}{Character vector. Vector con los identificadores internos de las estaciones.}

\item{pkvars}{Character vector. Vector con los identificadores internos de las variables.}

\item{pkfecs}{Character vector. Vector con las identificadores internos de las fechas.}
}
\value{
data.frame con columnas: pkest, pkvar, pkfec, valor.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la
Junta de Andalucía para obtener datos mensuales de varias estaciones, variables
y fechas, a partir de sus identificadores internos, y devolverlos como data.frame.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
pkests <- c('183') # EARM22
pkvars <- c('22') # TM1
pkfecs <- c('98981') # 31/12/2020
datos_mensuales <- getwsDatosMensualesRaw(idsesion,pkests,pkvars,pkfecs)
head(datos_mensuales)

}
