% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsEscalasList.R
\name{getwsEscalasList}
\alias{getwsEscalasList}
\title{Obtener la lista de las escalas temporales de la base de datos del CLIMA}
\usage{
getwsEscalasList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{CESCALA}, \code{DENOMINACION}, \code{PKESC}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de las escalas temporales en la base de datos CLIMA.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
escalas <- getwsEscalasList(idsesion)

}
