% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsEstacionesExpandidoList.R
\name{getwsEstacionesExpandidoList}
\alias{getwsEstacionesExpandidoList}
\title{Obtener lista detallada de estaciones meteorológicas del servicio CLIMA}
\usage{
getwsEstacionesExpandidoList(idsesion)
}
\arguments{
\item{idsesion}{\code{character}. Identificador de sesión válido obtenido mediante
\code{getwsIDSesion()}.}
}
\value{
Un \code{data.frame} con los principales atributos de cada estación, incluyendo
coordenadas, red, municipio, provincia, entorno y otros metadatos.
}
\description{
Esta función consulta el servicio web CLIMA de la Junta de Andalucía
y devuelve un \code{data.frame} con información detallada de todas las
estaciones meteorológicas disponibles en la base de datos.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
estaciones_detallado <- getwsEstacionesExpandidoList(idsesion)
head(estaciones_detallado)

}
