% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsEstacionesList.R
\name{getwsEstacionesList}
\alias{getwsEstacionesList}
\title{Obtener la lista de estaciones meteorológicas de la base de datos del CLIMA}
\usage{
getwsEstacionesList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{ACLPKACL}, \code{AGEPKAGE}, \code{CESTACION}, \code{COMPKCOM}, etcétera.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de estaciones meteorológicas de la base de datos CLIMA.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
estaciones <- getwsEstacionesList(idsesion)
head(estaciones)

}
