% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsEstadosDatosList.R
\name{getwsEstadosDatosList}
\alias{getwsEstadosDatosList}
\title{Obtener la lista de estados de los datos de la base de datos CLIMA}
\usage{
getwsEstadosDatosList(idsesion)
}
\arguments{
\item{idsesion}{\code{character}. Identificador de sesión válido obtenido mediante
\code{getwsIDSesion()}.}
}
\value{
Un \code{data.frame} con las columnas:
\itemize{
  \item \code{CTME} — Código del tipo de media.
  \item \code{DENOMINACION} — Nombre descriptivo del tipo de media.
  \item \code{PERIODO} — Periodo temporal de la media (por ejemplo, 24h).
  \item \code{PKTME} — Identificador único del tipo de media.
}
}
\description{
Esta función consulta el servicio web CLIMA de la Junta de Andalucía y devuelve un
\code{data.frame} con los estados de los datos.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
estados_datos <- getwsEstadosDatosList(idsesion)
head(estados_datos)

}
