% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsFechas.R
\name{getwsFechas}
\alias{getwsFechas}
\title{Obtener el identificador de una fecha de la base de datos del CLIMA}
\usage{
getwsFechas(date, idsesion)
}
\arguments{
\item{date}{fecha en formato DD/MM/YYYY}

\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{PKFEC}, \code{FECHA}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener el identificador de una fecha de la base de datos CLIMA.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
date <- "20/10/2025"
fecha <- getwsFechas(date,idsesion)
print(fecha)

}
