% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsFechasList.R
\name{getwsFechasList}
\alias{getwsFechasList}
\title{Obtener la lista de fechas de la base de datos del CLIMA}
\usage{
getwsFechasList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{CRED}, \code{DENOMINACION}, \code{OBSERVACION}, \code{PKRED}, \code{URL}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de las fechas en la base de datos CLIMA.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
fechas <- getwsFechasList(idsesion)
head(fechas)

}
