% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsGestoresList.R
\name{getwsGestoresList}
\alias{getwsGestoresList}
\title{Obtener la lista de gestores de redes meteorológicas de la base de datos del CLIMA}
\usage{
getwsGestoresList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{CGESTOR}, \code{DENOMINACION}, \code{PKGES}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de gestores de redes meteorológicas en la base de datos CLIMA.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
gestores <- getwsGestoresList(idsesion)
head(gestores)
}

}
