% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsIDSesion.R
\name{getwsIDSesion}
\alias{getwsIDSesion}
\title{Obtener ID de sesión del servicio de clima}
\usage{
getwsIDSesion(user, password)
}
\arguments{
\item{user}{Character. Nombre de usuario autorizado.}

\item{password}{Character. Contraseña correspondiente.}
}
\value{
Character. ID de sesión válido para usar en llamadas SOAP posteriores.
}
\description{
Esta función solicita un identificador de sesión (ID de sesión)
al servicio de clima de la Junta de Andalucía, autenticando con usuario y contraseña.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
print(idsesion)

}
