% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsMagnitudesList.R
\name{getwsMagnitudesList}
\alias{getwsMagnitudesList}
\title{Obtener la lista de magnitudes de la base de datos del CLIMA}
\usage{
getwsMagnitudesList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{CMAGNITUDES}, \code{DENOMINACION}, \code{PKMAG}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de magnitudes en la base de datos CLIMA.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
magnitudes <- getwsMagnitudesList(idsesion)
head(magnitudes)
}

}
