% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsMunicipiosList.R
\name{getwsMunicipiosList}
\alias{getwsMunicipiosList}
\title{Obtener la lista de municipios de la base de datos del CLIMA}
\usage{
getwsMunicipiosList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{PROVINCIA}, \code{MUNICIPIO}, \code{NOMBRE}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de municipios en la base de datos CLIMA.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
municipios <- getwsMunicipiosList(idsesion)

}
