% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsProvinciasList.R
\name{getwsProvinciasList}
\alias{getwsProvinciasList}
\title{Obtener la lista de provincias de la base de datos del CLIMA}
\usage{
getwsProvinciasList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{PROVINCIA}, \code{NOMBRE}, \code{PREFIJO_TEFN}, \code{COMUNIDAD}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de provincias en la base de datos CLIMA.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
provincias <- getwsProvinciasList(idsesion)
head(provincias)
}

}
