% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsRedesList.R
\name{getwsRedesList}
\alias{getwsRedesList}
\title{Obtener la lista de redes de estaciones meteorológicas (CLIMA)}
\usage{
getwsRedesList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{CRED}, \code{DENOMINACION}, \code{OBSERVACION}, \code{PKRED}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de redes disponibles en la base de datos CLIMA.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
redes <- getwsRedesList(idsesion)
head(redes)
}

}
