% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsUnidadesList.R
\name{getwsUnidadesList}
\alias{getwsUnidadesList}
\title{Obtener la lista de unidades de medida de la base de datos del CLIMA}
\usage{
getwsUnidadesList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{ABREVIATURA}, \code{DENOMINACION}, \code{PKUDM}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de unidades de medida en la base de datos CLIMA.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
unidades <- getwsUnidadesList(idsesion)

}
