% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsVariablesEstacionesList.R
\name{getwsVariablesEstacionesList}
\alias{getwsVariablesEstacionesList}
\title{Obtener la lista de estaciones asociadas a una variable a partir del
código de la variable y devolverla como data.frame}
\usage{
getwsVariablesEstacionesList(cvariable, idsesion)
}
\arguments{
\item{cvariable}{\code{character}. Identificador de la variable (CVARIABLE)
según el catálogo del servicio CLIMA.}

\item{idsesion}{\code{character}. Identificador de sesión válido proporcionado por
el servicio \code{getwsIDSesion()}.}
}
\value{
Un \code{data.frame} con las columnas:
\itemize{
  \item \code{ESTPKEST} — Identificador interno de la estación.
  \item \code{CESTACION} — Identificador de la estación.
  \item \code{EXISTENDATOS} — Indicador de disponibilidad de datos.
  \item \code{FACTOR} — Factor de corrección o calibración.
  \item \code{LIMITEVAL} — Límite de validación.
  \item \code{ORIGEN} — Origen de los datos.
  \item \code{PKVPE} — Identificador interno del vínculo variable-estación.
  \item \code{RANGOINFERIOR} — Rango inferior permitido para la variable.
  \item \code{RANGOSUPERIOR} — Rango superior permitido para la variable.
  \item \code{VARPKVAR} — Identificador interno de la variable
  \item \code{CVARIABLE} — Identificador de la variable.
}
}
\description{
Consulta el servicio web CLIMA de la Junta de Andalucía y devuelve un
\code{data.frame} con los códigos de las estaciones que registran datos para
una variable determinada.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
cvariable <- 'TI1'
variables_estaciones <- getwsVariablesEstacionesList(cvariable, idsesion)
print(variables_estaciones)
}

}
