% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsVariablesList.R
\name{getwsVariablesList}
\alias{getwsVariablesList}
\title{Obtener la lista de variables  meteorológicas de la base de datos del CLIMA}
\usage{
getwsVariablesList(idsesion)
}
\arguments{
\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{CVARIABLE}, \code{DENOMINACION}, \code{FRECUENCIA}, \code{PKVAR},  \code{UDMPKUDM}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener la lista de variables meteorológicas de la base de datos CLIMA.
}
\examples{
\dontrun{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
variables <- getwsVariablesList(idsesion)
head(variables)
}

}
