% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minutos_horas.R
\docType{data}
\name{minutos_horas}
\alias{minutos_horas}
\title{Tabla de equivalencia de minutos a horas}
\format{
Un \code{data.frame} con las siguientes columnas:
\describe{
  \item{minutos}{Número de minutos en el día, de 0 a 1440}
  \item{horas}{Hora en formato HH:MM}
}
}
\source{
Datos preprocesados internamente por el paquete.
}
\usage{
minutos_horas
}
\description{
Tabla de equivalencia de minutos a horas
}
\examples{
df <- minutos_horas
head(minutos_horas)
}
\keyword{datasets}
