% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsEstaciones.R
\name{getwsEstaciones}
\alias{getwsEstaciones}
\title{Obtener una estación de la lista de estaciones de la base de datos del CLIMA}
\usage{
getwsEstaciones(cestacion, idsesion)
}
\arguments{
\item{cestacion}{Character. Identificador de la estación.}

\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{ACLPKACL}, \code{AGEPKAGE}, \code{CESTACION}, \code{COMPKCOM}, etcétera.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener los datos de una estación meteorológica de la base de datos CLIMA.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
cestacion <- 'EARM22'
estacion <- getwsEstaciones(cestacion,idsesion)
head(estacion)

}
