% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwsVariables.R
\name{getwsVariables}
\alias{getwsVariables}
\title{Obtener una variable de la lista de variables  meteorológicas de la base de datos del CLIMA}
\usage{
getwsVariables(cvariable, idsesion)
}
\arguments{
\item{cvariable}{Character. Identificador de la variable.}

\item{idsesion}{Character. Identificador de sesión válido proporcionado por \code{getwsIDSesion}.}
}
\value{
data.frame con columnas: \code{CVARIABLE}, \code{DENOMINACION}, \code{FRECUENCIA}, \code{PKVAR},  \code{UDMPKUDM}.
}
\description{
Esta función realiza una llamada SOAP al servicio de clima de la Junta de Andalucía
para obtener una variable de la lista de variables meteorológicas de la base de datos CLIMA.
}
\examples{
user <- 'usuario'
password <- 'usuario'
idsesion <- getwsIDSesion(user,password)
cvariable <- 'TI1'
variable <- getwsVariables(cvariable,idsesion)
head(variable)

}
